+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.2] - 2026-08-23
+
+### Added
+- Reading input fields in mobile.php now show the last saved reading as placeholder
+
+### Fixed
+- Placeholder text color was not muted on iOS Safari for all mobile input fields
+
 ## [0.1.1] - 2026-06-07
 
 ### Added